rename sequenceScaner->directoryScanner; some efforts to unbork dpx input object type assignment
