+++ directoryScanner.py
